Update logic connect to redis

+++ MetroTicketBE.Application/SignalR/NotificationHub.cs
@@ -17,14 +17,19 @@ namespace MetroTicketBE.Domain.Entities
         public override async Task OnConnectedAsync()
         {
             var userId = Context.UserIdentifier;
-            await _redis.StoreKeyAsync($"checkLogin:{userId}", Context.ConnectionId);
+            await _redis.AddToSetAsync($"checkLogin:{userId}", Context.ConnectionId);
             await base.OnConnectedAsync();
         }
 
         public override async Task OnDisconnectedAsync(Exception? exception)
         {
             var userId = Context.UserIdentifier;
-            await _redis.DeleteKeyAsync($"checkLogin:{userId}");
+            await _redis.RemoveFromSetAsync($"checkLogin:{userId}", Context.ConnectionId);
+            var remaining = await _redis.GetSetCountAsync($"checkLogin:{userId}");
+            if (remaining == 0)
+            {
+                await _redis.DeleteKeyAsync($"checkLogin:{userId}");
+            }
             await base.OnDisconnectedAsync(exception);
         }
     }
